17029 FIX Synthetic monitoring: Tests with status NOT RUN, SKIP or NOT SET go stale
The test services become stale if the test status is NOT SET, NOT RUN, or SKIP.
Previously, tests with any of these statuses would go to the WARN state.
As a result of this werk, test services may change state and more the number of stale services may increase.

CMK-17658

Change-Id: I8e741f0981d5aad8904e61c817bf5fc9bfa3e385