EC: implement dunder ne for StructuredDataName
If you implement __eq__ but not __ne__, then == and != will behave somewhat strangely, since the default implementation of __ne__ is based on identity, not the negation of equality.
Quite often a class with only __eq__ will appear to work properly with !=, but it fails for various corner-cases (for example, an object which does not compare equal to itself, such as NaN).

Change-Id: I1c2ef9f9d331430e46dc0357c7d6e96c110d1027