Merge unit-resilience tests into unit-tests
It adds no value to have the randomized and the unrandomized unit tests.

Additionally, the benefit of running each test twice is close to zero.

Therefore this change merges the resilience tests into the default unit
tests and removes --count=2.

Also there is no distinction between xdist and sequential execution anymore,
as it adds no value and running the tests sequentially actually adds the risk
to break the parallelized tests without noticing.

Finally, this change replaces test-unit-slow with test-unit-all, which can
be run as part of the post-commit-chain.

Change-Id: Id3d5a3b4cc2d70d2c9aca7057ba94ad03dd10f6d