Rework HTML class
The previous API was sometimes misused and untrusted input was
converted into an HTML object. I guess one of the reasons is the tying
so if a function only accepts HTML and a dev needs to convert a str into
a HTML object that conversion was not obvious.

Since the HTML objects are nearly immutable we need to extend the
constructor. Making classmethods seems to be a good way to explicitly
state what should be done with the value.

CMK-17814

Change-Id: Iaa19365c4a6e760e73062cceed22bf17927ffb2c