ups_test: Handle unknown time since last test
The time since the last battery test is calculated via the difference between the current uptime and the uptime when the test was initiated.
This means the time cannot be determined if the uptime of the test initiation is larger than the current uptime.

Before, this lead to a
```
ValueError: Cannot render negative timespan
```

Now, the check will become UNKNOWN instead.

SUP-18881

Change-Id: I186adf739946923e52b844037887ed43bd2b6eab