Allow content negotiation in the REST API
The `accept` parameter, originally defined as a string can now also
be a list of strings indicating that more than one type of content is
acceptable by the endpoint.

In addition, the function is informed of the type of content received to facilitate processing.

Example:

```
@endpoint(
   ...
   accept=['application/json', 'application/gzip'],
)
def my_endpoint (params: Mapping[str, Any]) -> Result:
   body = params['body']
   content_type = params['content_type']
   ...
```

CMK-17757

Change-Id: Ib59eba172c6d9babdefdaf8fa1932b46305eda50