fix werk versions
There is no way that werks on the master branch can have 2.3.0 versions.

Only exception is 2.3.0b1. Those werks might be okay on the master
branch. This is because we currently add 2.4.0b1 werks to the master
branch.

In this change we adapt one werk that specified 2.3.0b1 as its version:
16559

But this werk was picked to the master branch, as there are two
commits for this werk: one on the master branch and one on 2.3.0, so its
correct to change this werks version to 2.4.0b1

Change-Id: I0ebabf493ca6b53bfbabcefe5342ef78b055b4d0