notify: create load_timeperiods func
The loading of timeperiods will be moved to watolib but
cmk.base.notify can't import from there.
We plan on passing the timeperiods as an arg everywhere
thats required in this module but this required a lot
more work. So for the time being, we will load from the
same module.

Change-Id: Ia8666dc087a99635d44af2b27417210ecd74f5e7