17089 SEC Change Transaction ID Format
More secure way to generate the transaction ID

The transaction ID is not intended as a security measure to protect us
from CSRF. Nonetheless, it sometimes saves us when we forgot to check
the CSRF token.
In the interest of best practices, generate it securely using the
secrets module, rather than random.

CMK-17244

Change-Id: I87b5e584b579ea4d5128143d25ad9a81de83bbba