16558 FIX Agent FreeBSD: Add missing Python interpreter detection
The Checkmk agent for FreeBSD, like other UNIX agents, handles the execution
of Python agent plugins by calling them with the available Python interpreter.

However, the functions for detecting the interpreter (or a missing one) were
recently missing, leading to a failure when trying to execute Python agent plugins.

Thanks to darkain for this contribution!

Closes #716

Change-Id: I5f23373fa272f230f468f945d84ed0ac83b49ad4