17001 FIX Enable several host actions no matter the tree depth of existing hosts
This Werk fixes the previous Werk #16638. For host actions "Run bulk service discovery", "Rename multiple hosts" and "Detect network parent hosts" only hosts in the current folder and in the first level subfolders were taken into account.
This is fixed to the expected recursive behavior: If a host exists in the current folder or any of its subfolders - no matter their tree depth - the host actions are enabled.

Change-Id: I52ad58854fa7e870d5b5fc9e8699040d7861a0f5