17075 Fix regex when combining multiple expressions using global flags
There are a few places where we concatenate multiple regular expressions
to save performance. This concatenation broke when expressions using
global flags such as "(?i)cpu" were concatenated. With this werk, we are
converting these global flags to local ones for that specific expression
which restores the expected behavior of the expression, fixing various
crashes that would occur for example when updating or activating
changes.

SUP-17283

Change-Id: Ib2719ce6bc7676388bc81a6e8fe95465f32d2ef2