17077 Allow filesystem service rule levels to go above 100%
We have one monster "filesystem" rule which is being used all over the
place. However, there have been quite a few support requests for single
checks coming in asking us to remove the 101% limit for the used space
level as they are monitoring virtualized file systems. However, making
just some checks use a specialized rule set of is a lot of work and in
the end it's easier to lift the restriction for the whole rule set, even
if some cases values above 100% still might not be valid.

Motivated by SUP-18744

Change-Id: I107a0c885ce03b2cfab7bcbfd2278896a040f3a4