Clean up dependency on Saas specific code
In general, the common Checmk code should not know anything about the
Saas specific code. In some cases this needs some extra work, which we
should always invest to keep a clean architecture. The edition specific
conditions are a way to fulfil the requirements quickly, but result in a
bad architecture which increase long term maintenance cost.

This change move Saas specific code to cse package and inverses the
dependency. The general Checkmk code now provides a hook which the Saas
code can use to extend the common Checkmk code . This way the code is
better encapsulated and the depdencies are clearer because we have less
ad-hoc extensions.

Change-Id: Ib1212ffb1c1c4f98bfbef3a8e222e8c0ba345f39