RMK: Keyword parsing: Use `messages` field instead of `message` field
The `messages` field seems to contain all messages. The `message` field
on the other hand contains only a single message and is empty if there
are only messages with logging level INFO.

This also fixes the issue that no messages were parsed for XMLs created
with Robot Framework version 6.

CMK-18043

Change-Id: Ib752f2097c056d3fad68e802c19250e6505f4abe