composition test: Cleanup fixtures and allow reusing site
For debug purposes, it is helpful to be able to rerun tests with
long setup phases, such as the composition tests. Therefore we
drop the test hosts at the end, so we can rerun
the tests without having to recreate the sites.

Additionally this change heavily reduces fixture usage by merging
and dropping many fixtures in an effort to make the tests more
understandable.

Change-Id: Icb27498b819aafe93ae54dbbe3d2e734c25caf2a