Introduce flask app for background jobs
The most straight forward way would be to use our make_wsgi_app helper
to construct a CheckmkApp as we do for regular requests. However, this
is not possible because of import cycles.

However, having worked a bit in that direction, it now looks like the
Checkmk specific app is not really needed. Instead we now build a
minimalistic job specific app, which seem to fit the needs we have in
the jobs.

Change-Id: If217f8701defc4442ac88deb2e140176e78ae2d7