Resolve cmk.gui dependency on background_job._process
The `_process` module needs to call `main_modules.load_plugins` which
would introduce circular dependencies with `cmk.gui`.

The `_process.run_process` function is the main entry point for the
subprocesses and as such not needed to be imported into `cmk.gui`.
Getting rid of all dependencies on `_process` of `cmk.gui` breaks the
cycle.

Change-Id: Iecb9f87adb71295c217c902e5575057afa3f8c4a