Change background jobs to use spawn instead of fork method
With the previous fork approach we traded fast job startup against
correctess and ultimately robustness. Doing that we took over part(!) of
the resources held by the origin process and had to clean up various
things and reinitialize some of them to have a sane execution
environment. However, we can not guarantee this for each and every
dependency we load into the UI code.

Since we are using threading in a few places and have the one or the
other native dependency, we can not be sure that the fork method does
not cause any trouble.

This change now changes the method to the spawn method. It does a fork,
followed by an exec which load a clean Python environment and executes
the target function.

This comes with some additional overhead when launching a background
job, but this seems to be acceptable for the jobs we currently have.

As a side effect we can now get rid of the special cleanup actions, like
disabling of the timeout manager, cleaning up of file handles and so on.
A few cleanups are done in this commit. Some follow in separate commits.

Change-Id: I165b0f439940d6aeac597de744689ca433c27257