kube_pod_conditions: remove summary
This functionality was broken since Werk 15588. `PodHasNetwork` is turned off
by default in all Kubernetes clusters, and does not exist in k8s v1.28 any
longer. This meant the check summary would depend on the feature gates enabled
by the customer.

There is no way to restore the old functionality, since we don't know whether
a condition is missing due the k8s version or the feature gates.

This change does not affect the monitoring capabilities, it merely
removes a distinct way to display the same data.

CMK-17818

Change-Id: I0da19d124ca8e65a2a46dd6fb4b4507d122c465f