REST API: Fix AttributeError when starting parent scan
Calling the "Start the parent scan background job" endpoint resulted in
the following exception:

'HostAddress' object has no attribute 'site_id'

Sadly the issue was not found by mypy because params is not well typed.
This is a systematical issue of the REST API we need to solve at some
point to get rid of this kind of issue.

Change-Id: I4409053ada83a1487f85a62ee77f7c1be225f045