17081 FIX Reduce risk of hanging processes
Currently files required for the licensing of Checkmk are locked for reading and writing. If a process does not release its acquired file lock for any reason, subsequent processes will continue to what for that lock, causing to the entire site to freeze.
This locking mechanism has now been removed to reduce the risk of this happening.

Change-Id: Ied775cffff0a935e8ce62adb6f2f259aacd55561