16562 FIX Fix automatic host removal in case one remote site is not logged in
The automatic host removal job is executed regularly in the background to remove
hosts from the monitoring once they cease to exist. In particular for but not
limited to automatically registered hosts.

This job failed in case one remote site was configured but not logged in, not
only affecting the not logged in site, but all sites. The not logged in site is
now being skipped, leaving the mechanism intact for all other sites.

Change-Id: I7f8a67eb8624354c76b74ba5ff7f43152c68b1bf