17144 Add/Increase required password length for backup/agent key generation
When generating a encryption key for backups previously the length of the passphrase was not checked, therefore `a` would have been a valid passphrase.

In the enterprise and cloud edition it is possible to generate a agent signing key where a passphrase is required.
Here the passphrase needed to be at minimum 8 characters.

With this Werk both passphrases need to be at least 12 characters long.

This only applies to newly generated keys.
Already existing keys with shorter passphrases are still valid.
Also if you upload a key-pair the length is not checked, since re-encrypting keys can be complicated and usually company policies already apply.

CMK-17535

Change-Id: I50c0f86b9dedc3344b5977e7e067ff38b359ac56