Don't show unused for some special agents
Previously, the `Effective Parameters` for a host would show `unused` in
some cases, and `no parameters` in order cases. With the `v1` ruleset
API only showing `no parameters` is possible. This is more consistent,
and also more descriptive.

This commit also removes some code for updating default values, which
was never executed.

CMK-17353

Change-Id: If8d6b88dd4e9d0a338b243b6c0091aec18dc08a9