Improve automatic host removal background job
At the end of the host removal background job an activation of changes
is triggered. Previously the job then finished without waiting for the
activation nor reporting about the result of the activation.

The job has now been changed to wait for the activation and report about
failed sites.

This also reduces flakes during the composition test runs.

Change-Id: Ia94c70405bf70a7a35eaea370eed84d60eb11c9b