Jaeger: Make Jaeger UI accessible through the site apache
We want an easy way to visualize the traces we have recorded in the
production environment. To do so, we ship Jaeger with Checkmk. This
change makes the Jaeger UI accessible through the site apache.

The Jaeger UI will be reached with "/[site]/jaeger" if TRACE_RECEIVE is
enabled via "omd config".

The authentication part is still missing and will be added in a
follow-up change.

CMK-18154

Change-Id: I4d12165727b76e4c54ba63f958c628dd42942851