Do not start activation cleanup job in case there is nothing to clean up
The background jobs now use the spawn mode instead of fork mode, which
makes them easier to manage correctly, but more expensive to start. In
the cases where we can determine that there is nothing to clean up, we
should not start the activation cleanup job at all to save resources.

Change-Id: Id70cc3533075309f621c3cf2c8fe7cbbba926c33