test-plugins-update: Skip dumps
In some situations there are so many checks that change from one branch
to another that would not make sense to skip them one by one.
For this reasong we here include the possibility to skip whole dumps
within the test-plugins-update.

In this way, we can re-enable the arista dumps for the test-plugins and
skip it for the test-plugins-update.

Change-Id: I079a567cf535ff318492ad72d458ebf92e7707ed