Fix background job startup timing issues
When the job inizializaion took more than 5 seconds, the job status
(get_status, is_active) could report a not running process while the
job is still being started.

This lead to various waiting routines, like waiting for a host renaming
background job via the REST API, to finish too early.

Details:

The startup roughly works like this:

1. Spawn a new python process.
2. Multiprocessing hands over to our run_process function.
3. The run_process function sets the *thread title* of the process.
4. The jobs business logic is being executed.

The thread title plays an important role in the identifcation of the
process to prevent issues with PID reuse.

When other processes ask for the state of the job, we check whether a
process with the same PID is still running. Then we check whether the
thread title is as expected to ensure that we are working with the right
process.

However, when the job is still being started as in 1. to 3. above, the
thread title is not yet set.

To handle this situation we had a 5 seconds grace period in place where
we would not check the thread title and assume the process to be the
right one. This logic does not work when the job initialization takes
longer than 5 seconds.

This change now tries to get rid of the timing issue by replacing the 5
second grace period with some kind of finger printing on the process
which is only applied in the initialization phase of the process. This
is not 100% bullet proof but is hopefully good enough for our use case.

CMK-18295

Change-Id: I2dd5536106006f847760624953568734cecd0942