Reland: Add initial implementation of trace module
It provides the basic functionality to setup tracing within a component,
adding logs as events to spans and instrumenting the code.

Also remove dependencies between cmk.trace and cmk.ccc.site which can
currently not depend on each other. Once cmk.ccc.site is moved to a
package or resolved in some other way, we can reconsider the right place
of TraceSendConfig and trace_send_config again.

CMK-18157

Change-Id: I4ee88894c248be26ddfa6ab88081066cb19db6f1