Refactor ExecutionMetrics class
The constructor of the previous approach did have two bugs, which are
fixed by this.
In the process of this the class has been transformed into a NamedTuple,
which should decrease the required memory for this a bit.

First setting the value of `bool` to an attribute led to this value
always being seen as `True` on a truthiness-check.
This could have led to reporting higher values for cache hits, cacheable
and remotable entries from the log.
These fields are now set `False` by default, which should report more
reliable results.

Second the check for the presence of `targetLabel` did not work as intended,
because it checked for the key being part of the string `"targetLabel"`
instead of checking if the key was the string `"targetLabel"`.
The check for an existing target label is now fixed.

Change-Id: I0b14f9018bb9800bbefb16b2a6cc31a97301da38