Improve UI initialization performance
It was observed that the initialization of the cmk.gui code executed
main_modules.load_plugins() twice. This was due to the fact that the the
UI and REST API blueprints were calling that function.

The research on this issue revealed that the Apache python environment
initialization, which is triggered through the mod_wsgi directive
WSGIImportScript was not triggering the main_modules loading as
intended.

This commit introduces a new import script that only cares about the
main modules imports. Since it is executed by mod_wsgi before requests
can be answered by apache processes, it should not slow down initial
requests anymore. Of course, the initial Apache process will need a bit
more time to be ready to handle requests now. But even this should be
better since the imports are done during startup and not during the first
request.

Change-Id: Ie8d18e8d262d297a0587d630ae2ec34407ea2a41