Resurrect linting via clang-tidy
We use bazel_clang_tidy for now, it looks very promising
due to the caching of the results.

Note that the clang-tidy configuration is now global again,
that's somehow the "Bazel way" of doing things. Perhaps the
bazel_clang_tidy upstream project might add something here,
but it's not really a big deal.

Change-Id: I91fe4bfa4961ab200a83509ca41add8c54716a0d