Consistently build all targets for native packages
Previously, "--build" did not build the tests, which was
inconsistent with e.g. the Rust packages.

Note that we still need to figure out how we want to handle
the --compilation_mode in general, the optimization flags,
-g/stripping, and NDEBUG. Currently we do a fastbuild with
flags for optimized code, we strip debug info, and we have
NDEBUG defined, which is a slightly weird combination.

Change-Id: Ibefaf4868c0de8227a1ad9bbedb25de737cc6fea