16887 FIX mk_job: MK_VARDIR defaults not being set in bakery
Due to a different way to set in `MK_VARDIR` in `mk_job`, default values would not be baked
into `mk_job` and derivates.
This change adds a replacement rule for the way `MK_VARDIR` gets assigned in `mk_job` and also
separates assignment and export in order to avoid known problems with Solaris.

Change-Id: I1bd53800e20cc73af768b842dfff197fb7071e08