Prevent API spec computation during tests
In production the REST API spec is automatically computed when the
tag configuration changes. So far, all tests which changed the tag
configuration triggered a background job and recomputed the REST API
spec, which is not necessary for the tests.

Besides increased and unnecessary load, this also lead to flakes in case
the job still ran during the test cleanup.

CMK-18469

Change-Id: If5c2a36218dee3dafa608cb07f6df3e992821a36