Prevent agent baking during tests
Adding hosts triggers an agent baking job in the background. The fixture
`suppress_bake_agents_in_background` is now used to prevent this. It
can be used for the following:

a) Skip calling the job
b) Ensure the job would have been called.

Of course, it would be great to do this without mocking, but this is way
beyond the scope of hot fixing the current situation in the tests which
start background jobs in an unintended way which leads to longer test
runs and even flakes in case jobs collide with test clean up logic.

CMK-18469

Change-Id: Iad66c08827b4e1518a06efe83228e1f525f04079