Fail all tests which start background jobs without annotation
It was just discovered that many tests execute background jobs which run
in a totally unsynchronized and unsupervised way. For example a test
which created a host, triggered the automatic agent baking in the
background. In general this is not the desired and in this example
background jobs should not be triggered at all, except the test is
specifically about the testing the background job.

Tests must not execute logic in background job processes, which
continue to run independently of the test case. To prevent that from
happening accidentally we now do the following:

1. Fail every test which starts a background job.
2. In case a test uses the fixture `inline_background_jobs`, the
   execution of the job is allowed.

Once annotated with the `inline_background_jobs` fixture, the background
job logic run synchronously so that the test code waits for the job to
complete. In many cases this is the desired behavior and makes it easier
to deal with the jobs in the tests.

However, in some cases you actually want to have a backrgound job being
executed in the same way as in the production environment. In that case,
you need to define a local fixture named "inline_background_jobs" to
override the global fixture.

CMK-18469

Change-Id: I8c47b82b90581fb2ab31d711dfe88731898ab91b