Quick hack to unblock CMK-18157
This is the minimalistic change to unblock CMK-18157. The missing RPATH
on the _message.so library breaks calls root calls of `omd` in the
moment we'd introduce the first tracing code to code used by `omd`.

A proper fix, considering all .so files built for python , will be
applied with CMK-18393.

Change-Id: I23061fa902a69fa723f3bb9148205e773ea7bf49