Reland: Introduce tracing to the UI WSGI app
This commit does the ground work for tracing in the UI.

It initializes the application wide tracing configuration and adds a
logger to attach log entries to spans.

Additionally we enable the auto instrumentation of Redis, requests and
the WSGI app and create a new span for each invocation of our WSGI app.

This is also the starting point to instrument the first distributed
program flow, which shall be: A view in distributed monitoring. The aim
is to get the data together from all involved components in a single
trace. Getting there will need some more work in the UI, livestatus
client, Liveproxy deamon and Livestatus which will be added step by
step.

CMK-18157

Change-Id: Ia55518b377e3d3959f6426a9596bb6b5776a3899