fix werk versions
you can quite easily check when a werk was actually released.
use the following command:

git log --oneline --graph --tags=v2.3.0* --branches=v2.3.0* --decorate

search for the werk number, and then you go up until you see the
first non rc release tag. if the tag is on a branch, also make sure
that the werk was picked to the regular 2.3.0 branch.

It's also not possible that 2.2.0 werks are on the 2.3.0 branch.

Change-Id: Ieedee1596438fcb7669c32e997be9897377ece91