17036 FIX heartbeat_crm: Handle cases when pacemaker service is not running
The creation of the agent section depended on the pacemaker service being running. If this was not the case, the section was left empty, causing the services to become stale.
From now on, if the pacemaker service is not running, the service will go into CRIT state and the summary will indicate that the connection was not possible/refused.

Also, the agent now checks for the existence of 'crm_mon' on the system, as this is a prerequisite for further command execution.

SUP-19573

Change-Id: I956f78411ffda122101b0076bc7ee643ee7983d4