16556 FIX Agent Updates: host selection ignores configured host labels
When configuring the global setting *Automatic agent updates/Activate update only on the selected hosts*,
the selection of host labels under *Match host labels* didn't get recognized.

Technical background: The set of host selection parameters used in above rule comes from a generic ruleset
pattern that's used in some more host rulesets in Checkmk.
Eventually, the option to filter for host labels got introduced to the generic ruleset, but we missed to
evaluate it for the determination of allowed hosts for agent updates.

SUP-19407
SUP-19585
SUP-19682

Change-Id: Ie279d31d55f8d5842c522a8fa8855191d1d2cd0a