Separate cmk.trace into modules
It turned out that the dependencies that are pulled in by cmk.trace
are too much for basic instrumenting use cases. The extreme is the OTLP
exporter, which pulls in protobuf. This is not an acceptable dependency
for code which just wants to create a span.

Luckily the issue was caused by own code. So the solution is fairly
straight forward: Separate `cmk.trace` into smaller parts which then can
be pulled in individually by the call sites as needed.

This is the first step which extracts the exporter setup and it's
dependencies.

Change-Id: Ib3791e4f9e5688bb923b0cf44d5d8f3be9eb6c51