17588 fritz: use 64-bit counters for traffic counters if available
Prior to this werk, the FRITZ!Box check plugin used traffic counters that are 32-bit wide.
Most current FRITZ!Box models also send additional 64-bit counters which should be used if
available to avoid the risk of overflowing counters. This werk adjusts the check plugin to
make use of those counters and fall back to the original 32-bit counters if they are unavailable.

Closes: #739

Change-Id: I3e4c75bd5539c233f35e2c1f562b6f65aef825c5