Fix bazel dependencies for cmk.trace
So far it seems that the dynamically generated bazel target
`trace_compile` (see omd/packages/python3-modules/BUILD) had a
dependency on the `packages/cmk-trace` directory. It looks like the
intention was to track all files below that directory, which is not how
it works.

This change tries to fix the broken dependencies by introducing a new
file group to cover all files of a package.

While this issue affects all `packages/cmk-*`, I won't touch it for
now. This will be done by other colleagues separately

Change-Id: Ibc0c81b60b9bca92510f6218f16852de3c2830ea