test-plugins-piggyback: Do not activate changes manually
After the DCD connector creates the piggyback-hosts, there is no need to
manually trigger a changes' activation for the created hosts.
The DCD logic will take care of it.

We here adapt the test to remove the manual changes' activation
and to wait for no-pending-changes to be present in the site after
the piggyback-hosts' creation.

Change-Id: Ibd276a8e14ef861ce024785e15fffb55b007b4ce