rabbitmq: Make distribution port configurable
RabbitMQ uses a "distribution port" for every node
to communicate with CLI tools (and other nodes).

In case the default port is in use (for example in
case we have multiple sites on one system), we must
be able to use a non-default port.

This port should not be opened to the netowrk.

CMK-18573

Change-Id: Ibd63426a8dae88985ce59b57b93cf0196aa937be