16866 FIX azure: Handle Azure API rate limit
After the changes in Azure API rate limits, the Azure agent requests to the API
were getting throttled for the big Azure environments.
This fix introduces handling for throttled requests. If the rate limit is reached,
the agent will repeat the request after 5 s. If it fails again, the agent will repeat the request
after another 10 s.

Additionally, the default limits for 'Lower levels for remaining API reads' in
the 'Azure Agent Info' monitoring rule are removed.

SUP-19155

Change-Id: I9df1be8d131e050d99b19dd4c631522ce01b0002