test-plugin-piggyback: Create new test module
We here separate test-plugin from test-plugin-piggyback in order to
make those two tests modules following a different assertion logic.

In test-plugin-piggyback we are not comparing the checks output with a
static reference anymore. Instead, we are asserting the following:

* All piggyback hosts retrieved from the test-site are exactly those
  present in the agent-dump;
* All services of the piggyback hosts are in 'OK' status.

The reason is that in the future this test will expand so that to use
dumps with hundreds of piggyback-hosts and thousands of services.
Then, it won't be feasible to store all checks output from all hosts.

CMK-18497

Change-Id: Id4ca43670f71d647dce6ce412d1a48aa1f2bb98f