composition test: Fix test_automatic_host_removal assertion
The assertion in the automatic_host_removal test was not very specific
and technically even wrong, because it assumed that no hosts have to
exist for the test to pass, while only two specific unresolvable test
hosts should be removed (any other hosts are irrelevant).

This change adds a more specific assertion to check only for the removal
of the unresolvable test hosts.

Change-Id: Ieb20d3d5de8b48953f00c9ebba3d05d2f5179839