core nagios: pass more values as arguments
This passes a few more configuration parameters as
arguments instead of directly accessing the global
variables.

In particular this includes the ones that are only
properly initialized if all plugins have been loaded.

Clearly this could be extended further, but we don't
need these options in the CMC case.

In particular: It would be worth while to find a way
that allows us to *not* load all the check plugins
in the CMC case; they are not needed there.

Change-Id: I94e7f1aff9720c66ed0db266c2a1e67c4eebb05c