Fix regression in fetch agent output job
The recent rework of the background job environment (fork vs. spawn)
results in no request and html globals being set. This is intended for
background jobs as they are running independent from a clients HTTP
request.

The job renders some HTML code, which now needs to be done through the
static HTMLGenerator.

Also add a test which would have caught this regression.

Change-Id: Ibd89e0a40118563a468405ae44554b9160893edf