16119 FIX Don't log unparsable SNMP traps per default
When an SNMP trap arrives at the event console which can't be parsed for
various reasons (unknown SNMP engine ID, some deserialization error, ...),
we previously logged a full Python backtrace to the EC log.  This could lead
to the incorrect conclusion that there is something wrong with the EC
itself, which is not what has actually happened: Actually, some device was
probably misconfigured and/or misbehaving.  To avoid such log spam, such a
trap is now silently dropped at the standard "informational" log level, and
more details can be logged at higher levels when needed.

Change-Id: I29e67adac4a10ea84df1089bb04567c639da217b