15339 FIX check_sql: activate thick mode for oracle connections
With [Werk #16023](https://checkmk.com/werk/16023) we switched the
library used to connect to oracle databases from `cx_Oracle` to
`oracledb`. For `cx_Oracle` it was mandatory to install "Oracle Instant
Client". The newer `oracledb` library has two modes: A stand alone thin
mode and a thick mode that needs the "Oracle Instant Client".

In order to fully replace `cx_Oracle`, `oracledb` needs to be in thick
mode. This is something we have not considered for Werk #16023.

With this Werk we now try to switch into the thick mode, and if this
does not work, we use thin mode. If you execute `check_sql` with `-v`
switch you will see a message if `oracledb` could not switch into thick
mode.

`oracledb` searches for the "Oracle Instant Client" in several standard
location. `check_sql` will find the installation if the files (among
other things `*.so` and `*.jar`) from the "Oracle Instant Client" files
are directly in `~local/lib/` in your site.

SUP-19387

Change-Id: I0349197f239b2cc18c15ee8f11c05a61044bad04