gui_e2e: add 'test_periodic_service_discovery_rule'
New test creates a new rule for periodic service discovery, moves it to
the top of the rules list and checks that the new rule was moved
successfully. Then navigates to the 'Effective parameters of host'
page and checks the presence of the new rule.

Three new POMs were created:
* AddRulePeriodicServiceDiscovery
* HostEffectiveParameters
* Ruleset

Ticket: CMK-17592
Change-Id: I6da43de43f14afc8fd07fff1e29c8cbf33b4bfe3