16867 FIX azure: Remove unnecessary 'metric not found' errors
When querying metrics, Azure agent was reporting each missing metric in the
details of the 'Azure Agent Info' service. This didn't influence the status
of the service but still implied an error.

It was misleading because some metrics aren't defined in some types
of resources. Now, those metrics are optional and won't cause an error.

SUP-19726

Change-Id: If8f370610c32252f0f9499c4bfed4a60a4910203